updated clientSocket.c and networkingTest.cpp
This is an a start at making the client manager. There is no threading yet, but it can handle multiple clients. Theres a few small problems such as occasionally not being able to bind to a port. This can be fixed by waiting for a bit I think. The next things which need doing are working out why it only writes 8 characters to the socket, actually clean up the resources used, try and replace all the c stuff with c++ stuff and make the comments more enlightening.