Improved clientManger
clientManager as been moved to its own cpp. The mysterious "unable to bind to port" bug still eludes me, I tried closing the port that the clientManager uses, but for some reason that didnt fix it. The clientManager now has 3 methods sans constructor/destructor: broadcast, whisper, getResponse (I wanted to call it listen, but thats the name of the function used to connect to ports). There are bits of hackyness, such as sending a carriage return to signal to the client that a response is expected.

Lets get ready to get thready